Update deployment architecture
