+++ backend/api.yaml
@@ -4,8 +4,3 @@ entrypoint: gunicorn -b :$PORT main:app --timeout 120
 
 includes:
   - env_variables.yaml
-
-instance_class: B4_1G
-
-basic_scaling:
-  max_instances: 4
