+++ backend/scraper.yaml
@@ -1,0 +1,11 @@
+service: scraper
+runtime: python37
+entrypoint: gunicorn -b :$PORT scraper_main:app --timeout 120
+
+includes:
+  - env_variables.yaml
+
+instance_class: B4_1G
+
+basic_scaling:
+  max_instances: 1
