+++ backend/scraper_cron/background_job/statistics_calculator.py
