+++ backend/scraper_cron/background_job/stock_frequency.py
