+++ backend/scraper_main.py
@@ -1,0 +1,1 @@
+from scraper_cron import app
