+++ cron.yaml
@@ -1,0 +1,5 @@
+cron:
+- description: "regular scraper job"
+  url: /scraper
+  schedule: every 30 minutes
+  target: scraper
